syn: remove `origin` information from errors
the idea here was to store the handle to the AST node related to the
origin (hence the name) of the error. but what really happens when an
errour occurs is that the node is not inserted in the AST at all -
meaning that the handle was completely bogus!

what we should do instead is to store a handle to the *token stream*,
not the *ast*, but thats work for later.