vm: use a fixed-length instruction encoding
by replacing the old, variable-sized, instruction enconding mechanism we
can represent the program as a chunk of `Op`, tying together opcodes and
their arguments, which *greatly* simplifies the readability of the vm.

operations are currently 8 bytes, which is pretty big. the sample
program under bin/file goes from 15 to 128 (!) byte. but it's worth it.