syn+vm: greatly improve the information associated with symbol decls
previously each symbol got an optional 'kind' and an optional 'expr',
which were supposed to represent the thing going in between the colon
and the equal sign, and the equal sign and the semicolon, respectively.

this was not enough to represent all the different symbols the language
will have.

now each symbol gets an obligatory, not optional, 'info' field, which
contains all the different combinations of information