vm+syn: properly implement scoping of symbols and function codegen
this is implemented using the concept of an "Enviroment", which is an
entity that maps identifiers -> symbols for the current lexical scope,
and an "Environment Manager", which is stores a collection of
environments that provides the actual api used by the bytecode generator
to search for and interact with the all symbols currently available in
the current lexical scope and all the ones above it.

this is very similar to how i implemented it in lox
(https://github.com/iniw/lox/blob/e1128e0798b0c6d7cdafb51f219839cc02dafa98/src/rt.rs#L844)
but much simpler, because pom doesn't have closures.