Sema: Half-finished implementation of function bindings type checking
This ultimately led to the realisation that functions don't
really align with other terms.

I'm designing a better way to declare function abstractions, which will
make them first class citizens and hopefully fix a lot of the weird stuff
in this patch, like the ordering of analysis when type checking function
bindings.

This new design is raising some questions about type inference though,
which I'm still not sure how to address.